databse connectivity and solve some bugs

+++ app_functions.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String, Date, Table, LargeBinary
+from sqlalchemy import Column, ForeignKey, Integer, String, Date, LargeBinary
 from sqlalchemy.orm import DeclarativeBase
 from sqlalchemy.orm import relationship
 from sqlalchemy import create_engine
@@ -6,8 +6,13 @@ from sqlalchemy.orm import sessionmaker
 import bcrypt, io
 from PIL import Image
 from datetime import datetime
-from sqlalchemy import select, delete
+from sqlalchemy import select
+import smtplib
 
+EMAIL_HOST = "smtp.gmail.com"
+EMAIL_HOST_USER = "azb1378@gmail.com"
+EMAIL_HOST_PASSWORD = "dkekzjnsjyzuabka"
+EMAIL_PORT_SSL = 465
 
 from load_saved_model import load_model
 model = load_model()
@@ -23,6 +28,7 @@ class Doctor(Base):
     email = Column(String(100), primary_key=True, nullable=False)
     password_hash = Column(Integer, nullable=False)
     name = Column(String(100), nullable=True)
+    bio = Column(String, nullable=True)
     patient_table = relationship("Patient", back_populates="doctor_table")
 
 class Patient(Base):
@@ -36,11 +42,6 @@ class Patient(Base):
     service_date = Column(Date, nullable=False)
     note = Column(String(100), nullable=False)
     doctor_table = relationship("Doctor", back_populates="patient_table")
-
-# class DoctorPatient(Base):
-#     __tablename__ = 'doctor_patient'
-#     doctor_email = Column(Integer, ForeignKey('doctors.email'), primary_key=True)
-#     patient_id = Column(Integer, ForeignKey('patients.id'), primary_key=True)
 
 
 # Create the database engine
@@ -76,7 +77,7 @@ def check_login(data):
     username = data["username"]
     input_password = data["pass"].encode("utf-8")
 
-    doctor_row = session.scalars(select(Doctor).where(Doctor.email.in_([username]))).one()
+    doctor_row = session.query(Doctor).filter_by(email=username).first()
 
     if (doctor_row):
 
@@ -97,29 +98,34 @@ def convert_bytes_image(image_bytes):
 def save_patient(datas):
     username = datas["u"]
     id_number = datas["i"]
-    mri_img = datas["m"]
+    mri_img = datas["m"].read()
     firstname = datas["f"]
     lastname = datas["l"]
-    note = datas["n"]
-    date = datetime.strptime(datas["d"], "%Y/%m/%d")
+    note = datas["n"].replace('\r', '\\r').replace('\n', '\\n')
+    if datas["d"] == "":
+        date = datetime.now()
+    else:
+        date = datetime.strptime(datas["d"], "%Y/%m/%d")
 
-    if mri_img != "":
+    if mri_img != b"":
         diagnose = model.predict2(convert_bytes_image(mri_img))
     else:
         diagnose = ""
     
     doctor_patient = session.scalars(select(Patient).filter_by(id=id_number).filter_by(doctor_email=username)).all()
 
+    ##### Create patient if it not exists
     if not doctor_patient:
         doctor = session.scalars(select(Doctor).where(Doctor.email.in_([username]))).one()
         patient = Patient(id=int(id_number), firstname=firstname, lastname=lastname, diagnose=diagnose, image_bytes=mri_img, service_date=date, note=note, doctor_table=doctor)
         session.add_all([patient])
 
+    ##### Edit patient if it exists
     else:
         doctor_patient[0].lastname = lastname
         doctor_patient[0].firstname = firstname
         doctor_patient[0].image_bytes = mri_img
-        doctor_patient[0].note = note
+        doctor_patient[0].note = note.replace('\r', '\\r').replace('\n', '\\n')
         doctor_patient[0].service_date = date
         doctor_patient[0].diagnose = diagnose
 
@@ -145,22 +151,110 @@ def get_patient(username, search_id=None):
     return contents
 
 
-def delete_patient(username, id):
+def delete_patient(username, id, just_delete_image=None):
     row_to_delete = session.query(Patient).filter_by(id=id).filter_by(doctor_email=username).first()
     if row_to_delete:
-        session.delete(row_to_delete)
+        if just_delete_image:
+            row_to_delete.image_bytes = b""
+            row_to_delete.diagnose = ""
+        else:
+            session.delete(row_to_delete)
         session.commit()
         return True
 
 
 def fast_diagnose(datas):
-    mri_img = datas["m"]
-    if mri_img != "":
+    mri_img = datas["m"].read()
+    if mri_img != b"":
         diagnose = model.predict2(convert_bytes_image(mri_img))
     else:
         diagnose = ""
     
     return diagnose
+
+
+def get_patient_image(username, id):
+    image_row = session.query(Patient).filter_by(id=id).filter_by(doctor_email=username).first()
+    if image_row:
+        image_bytes = image_row.image_bytes
+        return io.BytesIO(image_bytes)
+
+
+def modify_user_inf(datas):
+    username = datas["u"]
+    name = datas["n"]
+    bio = datas["b"].replace('\r', '\\r').replace('\n', '\\n')
+    input_password = datas["p"]
+    new_password = datas["np"]
+    new_email = datas["ne"]
+
+    user_or_pass_changed = False
+
+    doctor = session.query(Doctor).filter_by(email=username).first()
+    
+
+    if input_password and new_password:
+        saved_password_hash = doctor.password_hash
+        input_password = input_password.encode("utf-8")
+        if bcrypt.checkpw(input_password, saved_password_hash):
+            new_password_hash = bcrypt.hashpw(new_password.encode("utf-8"), bcrypt.gensalt())
+            doctor.password_hash = new_password_hash
+            mes_type = 1
+            user_or_pass_changed = True
+        else:
+            return -1
+
+    if new_email and new_email != username:
+        doctor_new_email = session.query(Doctor).filter_by(email=new_email).first()
+        if not doctor_new_email:
+            doctor.email = new_email
+            doctor_patient = session.query(Patient).filter(Patient.doctor_email==username).update({"doctor_email":new_email})
+            mes_type = 2
+            user_or_pass_changed = True
+        else:
+            return -2
+    
+    if not user_or_pass_changed:
+        mes_type = 0
+    
+    doctor.name = name
+    doctor.bio = bio
+    
+    session.commit()
+    return mes_type
+
+def get_user_inf(username):
+    doctor = session.query(Doctor).filter_by(email=username).first()
+    name = doctor.name 
+
+    bio = doctor.bio
+    bio = bio
+    contents={"u":username, "b":bio , "n":name}
+
+    return contents
+
+
+def send_mail(user_email, token):
+    
+
+    with smtplib.SMTP_SSL(EMAIL_HOST, EMAIL_PORT_SSL) as server:
+        server.login(EMAIL_HOST_USER, EMAIL_HOST_PASSWORD)
+        print(1)
+        server.sendmail(EMAIL_HOST_USER, user_email, )
+
+    
+
+
+
+            
+
+
+
+
+
+
+
+# print(get_patient_image("azb1378@gmail.com", "55"))
 
 
 
@@ -189,4 +283,4 @@ def fast_diagnose(datas):
 # save_patient(datas)
 
 # print(get_patient("ali@3516", search_id="23"))
-delete_patient("ali@3516", "356")
+# delete_patient("ali@3516", "356")
